refactor: Consolidate `PromptCancelError` into `entry-util.ts` & remove `prompt-cancel.ts`
