--- src/prompt-cancel.ts
@@ -1,6 +1,0 @@
-export class PromptCancelError extends Error {
-  constructor() {
-    super('Prompt canceled');
-    this.name = 'PromptCancelError';
-  }
-}
